Add authorization to BooksController#[index,show]
Add basic authorization to the BooksController.

This is basically a stub until actual signup/signin is implemented.
Currently there is no way (outside of specs) to log in a user (i.e.
actually get the necessary token), nor tho signup (short of adding a
user manually in the Rails console or through a spec)

The implementation and specs ensure that the actual authorization pieces
are working, so we can concentrate on the signup/signin process, then
work on adding authorization to everting else that needs it.