Add SigninController
Add controller actions for signing up. This is very similar to the
signup controller in the reponse format, with the exception that it is
validating an existing user instead of creating a new one.

As with signups, the reponse includes a csrf token in the JSON body and
a jwt_access token in the cookies header