Add OpenlibraryApi service adapter
As a first step to implementing ISBN lookup, we add an adapter to use
the OpenLibrary API (a free and open API that doesn't need API keys to
access it),  and then parse it's output for the data we need.

Because we're testing external APIs for the first time, we add and
configure the VCR gem here as well.