Add lookup_isbn endpoint
Add an endpoint '/api/v1/lookup_isbn' for looking up isbn from an
external API and returning a JSON object to fill out a form for a new
Book entry.

Of interest is how the `author` attribute is treated. When the Author
already exists in the database, the `author` attribute is returned as
`{id: author.id, name: author.name}`. However, if this is a "new"
Author, the `id` is returned as `-1`. This is meant to signify to a
front-end application that a new Author should be created if the lookup
is being used to fill in a form for a new Book.