Migrate Book.year to type :date and update Lookup
This commit updates the (publication) year field for Books to a Date
format, and normalized the output of the ISBN lookup for said field.

The Lookup output needs to be normalized to a format that will be used
by the front-end's date-picker. We've chosen ISO 8601 standard YYYY-MM-DD
format for convenience and to eliminate ambiguity that might result from
MM-DD-YYYY or DD-MM-YYYY.

When the Openlibrary API returns only a year, the Date.parse would
normally be invalid. In those cases, we set the month and day to 01/01
(January 1st) and retry parsing the date, eg `Date.parse("#{year}/01/01")`.